chore: add dashboard to shared models
