+++ src/shared/models/dashboard.ts
@@ -1,0 +1,10 @@
+export type DashboardState = {
+  isOpen: boolean;
+  ui: {
+    activeMenu: string;
+    selectedNamespace: string;
+  };
+  tableDrawer: {
+    selectedResourceId?: string;
+  };
+};
